refactor: Simplified shared code
