+++ packages/eslint/ember/v2-addon/index.mjs
@@ -10,7 +10,7 @@ import eslintPluginTypescriptSortKeys from 'eslint-plugin-typescript-sort-keys';
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
 
-import customRules from '../custom-rules/index.mjs';
+import { customRules } from '../shared/index.mjs';
 
 const parserOptionsJs = {
   ecmaFeatures: {
